add critical calc add rank calc etc...

+++ BattleInfoPlugin/Models/Raw/CommonTypeExtensions.cs
@@ -118,7 +118,7 @@ namespace BattleInfoPlugin.Models.Raw
         /// </summary>
         /// <param name="damages">api_fdam/api_edam</param>
         /// <returns></returns>
-        public static FleetDamages GetDamages(this double[] damages)
+        public static FleetDamages GetDamages(this decimal[] damages)
             => damages
                 .GetFriendData() //敵味方共通
                 .Select(Convert.ToInt32)
